Fixed a bug where the save data for the console was not being stored on tab restart.
We now have a new method in the Tab bas class called `get_all_save_data()` which collates the save data from `Tab.get_builtin_save_data` and the `get_save_data` method of a subclass (if it exists).
We also fix a bug which causes blacs to crash on exit if the return value of `get_save_data()` was not a dictionary.
We also change the order in which the dictionary gets updated so that subclasses can override the restore behaviour of the worker terminal if they wish (but it's discouraged via the use of the underscore prepending the dictionary keys).
The front panel settings saving code is also now simplified to make use of these changes!

This fixes issue #47 and results in a bump of the patch version to 2.3.1